refactor: serde trait file name
