+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/create_tree.rs
@@ -13,8 +13,8 @@ use crate::patricia_merkle_tree::types::TreeHeight;
 use crate::patricia_merkle_tree::{
     original_skeleton_tree::node::OriginalSkeletonNode, types::NodeIndex,
 };
+use crate::storage::db_object::Deserializable;
 use crate::storage::errors::StorageError;
-use crate::storage::serde_trait::Deserializable;
 use crate::storage::storage_trait::create_db_key;
 use crate::storage::storage_trait::Storage;
 use crate::storage::storage_trait::StorageKey;
