+++ crates/committer/src/storage.rs
@@ -1,4 +1,4 @@
+pub mod db_object;
 pub mod errors;
 pub mod map_storage;
-pub mod serde_trait;
 pub mod storage_trait;
